functions/download_dados_farm/requirements.txt
functions-framework==3.*
requests==2.*
pandas==2.*
pyarrow==14.*
google-cloud-storage==2.*